services/user_service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = "user service"
authors = ["emikeb <no@email.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-consul = "^1.1.0"
flask = "^3.0.3"
coverage = "^7.6.0"
pika = "^1.3.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
